tidy up old files
