--- buildx.sh
@@ -1,1 +1,0 @@
-docker buildx build --platform linux/amd64,linux/arm64,linux/arm/v7,linux/arm/v6 -t emellodew/giv_tcp:dev --push .
