--- givenergy_modbus/__init__.py
@@ -1,5 +1,0 @@
-"""Top-level package for GivEnergy Modbus."""
-
-__author__ = """Dewet Diener"""
-__email__ = 'givenergy-modbus@dewet.org'
-__version__ = '0.10.1'
